Move processing graph functions into a separate API
