+++ lib/engine_api/api/project_api.dart
@@ -62,129 +62,6 @@ class ProjectApi {
     _engine._request(id, request);
   }
 
-  Future<List<ProcessorDescription>> getProcessors() {
-    final completer = Completer<List<ProcessorDescription>>();
-
-    final id = _engine._getRequestId();
-
-    final request = RequestObjectBuilder(
-      id: id,
-      commandType: CommandTypeId.GetProcessors,
-      command: GetProcessorsObjectBuilder(),
-    );
-
-    _engine._request(id, request, onResponse: (response) {
-      final inner = response.returnValue as GetProcessorsResponse;
-      completer.complete(inner.processors!);
-    });
-
-    return completer.future;
-  }
-
-  Future<int> getMasterOutputNodeId() {
-    final completer = Completer<int>();
-
-    final id = _engine._getRequestId();
-
-    final request = RequestObjectBuilder(
-      id: id,
-      commandType: CommandTypeId.GetMasterOutputNodeId,
-      command: GetMasterOutputNodeIdObjectBuilder(),
-    );
-
-    _engine._request(id, request, onResponse: (response) {
-      final inner = response.returnValue as GetMasterOutputNodeIdResponse;
-      completer.complete(inner.nodeId);
-    });
-
-    return completer.future;
-  }
-
-  /// Adds the processor with the given ID to the engine's processing graph.
-  Future<int> addProcessor(String processorId) {
-    final id = _engine._getRequestId();
-
-    final request = RequestObjectBuilder(
-      id: id,
-      commandType: CommandTypeId.AddProcessor,
-      command: AddProcessorObjectBuilder(
-        id: processorId,
-      ),
-    );
-
-    final completer = Completer<int>();
-
-    _engine._request(id, request, onResponse: (response) {
-      final inner = response.returnValue as AddProcessorResponse;
-      if (inner.success) {
-        completer.complete(inner.processorId);
-      } else {
-        completer.completeError(false);
-      }
-    });
-
-    return completer.future;
-  }
-
-  /// Connects two processors in the engine's processing graph.
-  Future<void> connectProcessors({
-    required int sourceId,
-    required int sourcePortIndex,
-    required int destinationId,
-    required int destinationPortIndex,
-    required ProcessorConnectionType connectionType,
-  }) {
-    final completer = Completer<void>();
-
-    final id = _engine._getRequestId();
-
-    final request = RequestObjectBuilder(
-      id: id,
-      commandType: CommandTypeId.ConnectProcessors,
-      command: ConnectProcessorsObjectBuilder(
-        sourceId: sourceId,
-        sourcePortIndex: sourcePortIndex,
-        destinationId: destinationId,
-        destinationPortIndex: destinationPortIndex,
-        connectionType: connectionType,
-      ),
-    );
-
-    _engine._request(id, request, onResponse: (response) {
-      if ((response.returnValue as ConnectProcessorsResponse).success) {
-        completer.complete();
-      } else {
-        completer.completeError(false);
-      }
-    });
-
-    return completer.future;
-  }
-
-  Future<void> compileProcessingGraph() {
-    final completer = Completer<void>();
-
-    final id = _engine._getRequestId();
-
-    final request = RequestObjectBuilder(
-      id: id,
-      commandType: CommandTypeId.CompileProcessingGraph,
-      command: CompileProcessingGraphObjectBuilder(),
-    );
-
-    _engine._request(id, request, onResponse: (response) {
-      if ((response.returnValue as CompileProcessingGraphResponse).success) {
-        completer.complete();
-      } else {
-        completer.completeError(
-            (response.returnValue as CompileProcessingGraphResponse).error ??
-                'Unknown error');
-      }
-    });
-
-    return completer.future;
-  }
-
   void noteOn({
     int channel = 1,
     required int note,
